--- a/lib/filter-builds.js
+++ b/lib/filter-builds.js
@@ -35,17 +35,13 @@ function hasDependencies(build) {
 }
 
 /**
- * Returns `true` the build in queue and has compatible agents.
+ * Returns `true` the build has compatible agents.
  *
  * @param {object} build - the build info.
  * @returns {boolean}
  */
-function isQueuedWithCompatibleAgents(build) {
-    const compatibleAgents = build.compatibleAgents;
-    const isQueued = build.state === 'queued';
-    const hasCompatibleAgents = compatibleAgents && compatibleAgents.length !== 0;
-
-    return isQueued ? hasCompatibleAgents : true;
+function hasCompatibleAgents(build) {
+    return build.waitReason !== 'This build will not start because there are no compatible agents which can run it';
 }
 
 /**
@@ -64,7 +60,7 @@ module.exports = (builds, options) => {
     return builds.filter(build => {
         if (projectPattern && !isMatchProjectName(build, projectPattern)
             || options.ignoreDependencies && hasDependencies(build)
-            || options.ignoreIncompatibleAgents && !isQueuedWithCompatibleAgents(build)
+            || options.ignoreIncompatibleAgents && !hasCompatibleAgents(build)
         ) {
             return false;
         }