--- a/lib/filter-builds.js
+++ b/lib/filter-builds.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const mm = require('micromatch');
 
 /**