--- a/test/api.test.js
+++ b/test/api.test.js
@@ -5,21 +5,19 @@ const proxyquire = require('proxyquire');
 const sinon = require('sinon');
 
 test.beforeEach(t => {
-    t.context.gotStub = sinon.stub().resolves({ body: {} });
     t.context.loadStub = sinon.stub().resolves([]);
     t.context.filterStub = sinon.stub().returns([]);
 
     t.context.run = proxyquire('../index', {
-        got: t.context.gotStub,
-        './lib/load-build': t.context.loadStub,
+        './lib/load-builds': t.context.loadStub,
         './lib/filter-builds': t.context.filterStub
     });
 });
 
-test('should get info by url', async t => {
+test('should load builds of queue', async t => {
     await t.context.run('http://tc.url');
 
-    t.true(t.context.gotStub.calledWith('http://tc.url/guestAuth/app/rest/buildQueue'));
+    t.true(t.context.loadStub.called);
 });
 
 test('should returns builds of queue', async t => {
@@ -40,18 +38,4 @@ test('should returns queue size', async t => {
     const queue = await t.context.run('http://tc.url');
 
     t.is(queue.size, builds.length);
-});
-
-test('should load build info', async t => {
-    t.context.gotStub.resolves({
-        body: {
-            href: '/guestAuth/app/rest/buildQueue',
-            count: 1,
-            build: [{ href: 'build-url-1' }]
-        }
-    });
-
-    await t.context.run('http://tc.url');
-
-    t.true(t.context.loadStub.calledWith('http://tc.url/build-url-1'));
 });