--- a/test/load-json.test.js
+++ b/test/load-json.test.js
@@ -1,0 +1,25 @@
+'use strict';
+
+const test = require('ava');
+const proxyquire = require('proxyquire');
+const sinon = require('sinon');
+
+test.beforeEach(t => {
+    t.context.gotStub = sinon.stub().resolves({ body: { res: 'bla' } });
+
+    t.context.load = proxyquire('../lib/load-json', {
+        got: t.context.gotStub
+    });
+});
+
+test('should return json', async t => {
+    const info = await t.context.load('url');
+
+    t.deepEqual(info, { res: 'bla' });
+});
+
+test('should throw error', t => {
+    t.context.gotStub.rejects('err');
+
+    t.throws(t.context.load('url'), 'err');
+});