We can use reason status instead of load info about agents.

Use debug to log info about http requests.

**Log example:**

```
  http loading build queue list +0ms
  http build queue list successfully loaded in 218ms +218ms
  http loading build 3318837 +1ms
  http loading build 3318782 +2ms
  http loading build 3318785 +0ms
  http loading build 3306642 +1ms
  http loading build 3244874 +0ms
  http loading build 3244875 +1ms
  http loading build 3274775 +0ms
  http loading build 3256326 +0ms
  http loading build 3318768 +1ms
  http loading build 3318769 +0ms
  http loading build 3318771 +1ms
  http loading build 3318770 +0ms
  http loading build 3318772 +1ms
  http loading build 3318832 +0ms
  http loading build 3318841 +1ms
  http loading build 3318840 +1ms
  http loading build 3318844 +4ms
  http loading build 3318846 +1ms
  http loading build 3318847 +1ms
  http loading build 3318851 +0ms
  http loading build 3318854 +1ms
  http loading build 3318856 +0ms
  http loading build 3318858 +1ms
  http loading build 3318861 +1ms
  http loading build 3318863 +0ms
  http loading build 3318864 +1ms
  http loading build 3318870 +2ms
  http loading build 3318865 +1ms
  http loading build 3318866 +0ms
  http loading build 3318867 +1ms
  http loading build 3318868 +1ms
  http loading build 3318869 +1ms
  http loading build 3318871 +1ms
  http loading build 3318874 +0ms
  http loading build 3318877 +1ms
  http loading build 3318880 +0ms
  http loading build 3318883 +1ms
  http loading build 3318885 +0ms
  http loading build 3318887 +0ms
  http loading build 3318889 +1ms
  http loading build 3318891 +0ms
  http loading build 3318893 +0ms
  http loading build 3318895 +1ms
  http loading build 3318896 +0ms
  http loading build 3318898 +1ms
  http loading build 3318901 +0ms
  http loading build 3318903 +1ms
  http loading build 3318905 +1ms
  http loading build 3318907 +0ms
  http loading build 3318909 +0ms
  http loading build 3318911 +1ms
  http loading build 3318913 +0ms
  http loading build 3318917 +0ms
  http loading build 3318923 +1ms
  http loading build 3318926 +0ms
  http loading build 3318929 +0ms
  http loading build 3318932 +1ms
  http loading build 3318935 +0ms
  http loading build 3318938 +0ms
  http loading build 3318941 +0ms
  http loading build 3318944 +1ms
  http loading build 3318945 +0ms
  http loading build 3318946 +1ms
  http loading build 3318947 +0ms
  http loading build 3318949 +1ms
  http loading build 3318950 +2ms
  http loading build 3318953 +0ms
  http loading build 3318955 +1ms
  http loading build 3318957 +0ms
  http loading build 3318958 +0ms
  http loading build 3318959 +0ms
  http loading build 3318960 +1ms
  http loading build 3318961 +0ms
  http loading build 3318963 +1ms
  http loading build 3318965 +0ms
  http loading build 3318966 +0ms
  http loading build 3318969 +1ms
  http loading build 3318972 +0ms
  http loading build 3318975 +1ms
  http loading build 3318978 +0ms
  http loading build 3318981 +0ms
  http loading build 3318984 +1ms
  http loading build 3318987 +0ms
  http loading build 3318990 +1ms
  http loading build 3318991 +1ms
  http loading build 3318349 +0ms
  http build 3244874 successfully loaded in 92ms +46ms
  http build 3306642 successfully loaded in 93ms +1ms
  http build 3318837 successfully loaded in 102ms +6ms
  http build 3244875 successfully loaded in 99ms +1ms
  http build 3318785 successfully loaded in 101ms +0ms
  http build 3318769 successfully loaded in 108ms +10ms
  http build 3318771 successfully loaded in 117ms +10ms
  http build 3318840 successfully loaded in 115ms +1ms
  http build 3256326 successfully loaded in 130ms +10ms
  http build 3274775 successfully loaded in 131ms +1ms
  http build 3318770 successfully loaded in 129ms +0ms
  http build 3318772 successfully loaded in 136ms +8ms
  http build 3318864 successfully loaded in 124ms +0ms
  http build 3318847 successfully loaded in 139ms +11ms
  http build 3318856 successfully loaded in 274ms +136ms
  http build 3318861 successfully loaded in 273ms +1ms
  http build 3318870 successfully loaded in 272ms +2ms
  http build 3318858 successfully loaded in 277ms +1ms
  http build 3318869 successfully loaded in 269ms +0ms
  http build 3318885 successfully loaded in 272ms +6ms
  http build 3318865 successfully loaded in 279ms +1ms
  http build 3318768 successfully loaded in 299ms +3ms
  http build 3318866 successfully loaded in 282ms +0ms
  http build 3318782 successfully loaded in 308ms +6ms
  http build 3318868 successfully loaded in 291ms +5ms
  http build 3318905 successfully loaded in 283ms +1ms
  http build 3318871 successfully loaded in 294ms +4ms
  http build 3318891 successfully loaded in 291ms +0ms
  http build 3318880 successfully loaded in 296ms +3ms
  http build 3318863 successfully loaded in 306ms +1ms
  http build 3318832 successfully loaded in 322ms +5ms
  http build 3318895 successfully loaded in 304ms +5ms
  http build 3318911 successfully loaded in 301ms +1ms
  http build 3318946 successfully loaded in 301ms +4ms
  http build 3318935 successfully loaded in 303ms +0ms
  http build 3318929 successfully loaded in 309ms +5ms
  http build 3318896 successfully loaded in 316ms +2ms
  http build 3318926 successfully loaded in 311ms +0ms
  http build 3318887 successfully loaded in 321ms +3ms
  http build 3318938 successfully loaded in 533ms +220ms
  http build 3318883 successfully loaded in 543ms +3ms
  http build 3318893 successfully loaded in 543ms +0ms
  http build 3318889 successfully loaded in 545ms +2ms
  http build 3318877 successfully loaded in 552ms +5ms
  http build 3318903 successfully loaded in 549ms +2ms
  http build 3318844 successfully loaded in 570ms +2ms
  http build 3318898 successfully loaded in 552ms +0ms
  http build 3318851 successfully loaded in 568ms +0ms
  http build 3318932 successfully loaded in 553ms +6ms
  http build 3318950 successfully loaded in 549ms +1ms
  http build 3318913 successfully loaded in 556ms +0ms
  http build 3318917 successfully loaded in 557ms +1ms
  http build 3318941 successfully loaded in 813ms +258ms
  http build 3318909 successfully loaded in 817ms +1ms
  http build 3318953 successfully loaded in 810ms +1ms
  http build 3318923 successfully loaded in 817ms +1ms
  http build 3318947 successfully loaded in 814ms +0ms
  http build 3318949 successfully loaded in 1s +518ms
  http build 3318957 successfully loaded in 1s +2ms
  http build 3318854 successfully loaded in 1s +1ms
  http build 3318959 successfully loaded in 1s +0ms
  http build 3318960 successfully loaded in 1s +1ms
  http build 3318867 successfully loaded in 1s +6ms
  http build 3318901 successfully loaded in 1s +1ms
  http build 3318944 successfully loaded in 1s +1ms
  http build 3318955 successfully loaded in 1s +14ms
  http build 3318349 successfully loaded in 1s +1ms
  http build 3318991 successfully loaded in 1s +2ms
  http build 3318972 successfully loaded in 1s +0ms
  http build 3318969 successfully loaded in 1s +0ms
  http build 3318966 successfully loaded in 1s +0ms
  http build 3318958 successfully loaded in 1s +6ms
  http build 3318846 successfully loaded in 1s +8ms
  http build 3318945 successfully loaded in 1s +4ms
  http build 3318961 successfully loaded in 1s +1ms
  http build 3318984 successfully loaded in 1s +1ms
  http build 3318990 successfully loaded in 1s +2ms
  http build 3318965 successfully loaded in 1s +0ms
  http build 3318981 successfully loaded in 1s +2ms
  http build 3318975 successfully loaded in 1s +5ms
  http build 3318907 successfully loaded in 1s +1ms
  http build 3318874 successfully loaded in 1s +5ms
  http build 3318978 successfully loaded in 1s +0ms
  http build 3318987 successfully loaded in 2s +195ms
  http build 3318963 successfully loaded in 2s +13ms
  http build 3318841 successfully loaded in 3s +1s
  http build queue loaded in 3s +1ms
```