@anton-rudeshko 2016-06-04
Кстати, глядя на пример твоего лога, кажется странным, что каждый следующий билд фетчится медленнее:
```
http build 3244874 successfully loaded in 92ms
…
http build 3318858 successfully loaded in 277ms
…
http build 3318911 successfully loaded in 301ms
…
http build 3318932 successfully loaded in 553ms
…
http build 3318947 successfully loaded in 814ms
```
Или это ошибка при логгировании?