@blond 2016-06-06
В логгировании ошибки не вижу.

Могу предположить что TC не может отвечать на всё сразу. Например, внутри очередь на обращение к БД.